+++ src/routes/v2/admin/tasks/+page.server.js
@@ -1,13 +1,16 @@
-import { orderTasksByOwnerThenByNameThenByVersion } from '$lib/common/component_utilities';
-import { listLegacyTasks } from '$lib/server/api/v2/task_api.js';
+import { listUsers } from '$lib/server/api/v1/auth_api.js';
 
 export async function load({ fetch }) {
-	/** @type {import('$lib/types').Task[]} */
-	const tasks = await listLegacyTasks(fetch);
+	/** @type {import('$lib/types').User[]} */
+	const usersList = await listUsers(fetch);
 
-	orderTasksByOwnerThenByNameThenByVersion(tasks, null, 'desc');
+	const users = /** @type {string[]} */ (
+		usersList.map((u) => (u.username ? u.username : u.slurm_user)).filter((u) => !!u)
+	);
+
+	users.sort((a, b) => a.localeCompare(b, undefined, { sensitivity: 'base' }));
 
 	return {
-		tasks
+		users
 	};
 }